+++ packages/block-library/src/navigation-link/shared/test/use-handle-link-change.test.js
@@ -1,8 +1,4 @@
 /**
- * @jest-environment jsdom
- */
-
-/**
  * External dependencies
  */
 import { renderHook } from '@testing-library/react';