+++ packages/components/src/query-controls/category-select.tsx
@@ -29,7 +29,6 @@ export default function CategorySelect( {
 					: undefined
 			}
 			{ ...props }
-			__next40pxDefaultSize
 		/>
 	);
 }